Fix message flag bug
Currently, if you don't have a noti config file, setting the
notification message with --message doesn't work. This is because noti
stops configuration after reading the config file fails.

This change just logs that error instead of returning, since it's a
non-fatal error. It also fixes the unit test that should have caught
this.

Resolves #72