Fix double noti in Slack
Currently, the Slack notification combines the title and message to use for
Slack.Text. Also, the default username for the notification is noti. This makes
it so noti gets printed twice.

  username: noti
  text: noti
        hello

This change makes it so that if the title and the username are the same, then
only the message gets used as Slack.Text.

  username: noti
  text: hello

  username: mycompany
  text: noti
        hello