Add logic to read message from stdin with `--message=-` (#112)
This change supports using noti as the end of a pipeline of commands.
Reading from standard input adds flexibility.

Say you had a deployment script that outputs an artifact path, some job
statistics, and/or url for a canary instance. You could read that output
and notify people when & where the deploy is available.

See the updated documentation for more.